+++ src/operations.h
@@ -12,44 +12,44 @@ const size_t MEMORY_SPACE = 100; // space in memory; 100 words available
 //  Code = 10
 //  read() => reads a word from keyboard into passed in location 
 //  in memory to store  
-void read(array<int, MEMORY_SPACE>& mem, int& loc);
+void read(array<double, MEMORY_SPACE>& mem, int& loc);
 
 //  Code = 11
 //  write() => writes the word location, loc, in memory to the screen
-void write(array<int, MEMORY_SPACE>& mem, int& loc);
+void write(array<double, MEMORY_SPACE>& mem, int& loc);
 
 
 
 /*  Load/Store Operations   */
 //  Code = 20
 //  load() => load a word from location in memory into accumulator
-void load(array<int, MEMORY_SPACE>& mem, int& loc, long double& accum);
+void load(array<double, MEMORY_SPACE>& mem, int& loc, long double& accum);
 
 //  Code = 21
 //  store() => stores a word from the accumulator into a specific location in memory
-void store(array<int, MEMORY_SPACE>& mem, int& loc, long double& accum);
+void store(array<double, MEMORY_SPACE>& mem, int& loc, long double& accum);
 
 
 /*  Arithmetic Operations   */
 //  Code = 30
 //  add() => add a word from location in memory to the word in the
 //  accumulator (leaving result in accumulator)
-void add(array<int, MEMORY_SPACE>& mem, int& loc, long double& accum);
+void add(array<double, MEMORY_SPACE>& mem, int& loc, long double& accum);
 
 //  Code = 31
 //  subtract() => subtract a word from location in memory from the
 //  word in accumulator (leaving result in accumulator)
-void subtract(array<int, MEMORY_SPACE>& mem, int& loc, long double& accum);
+void subtract(array<double, MEMORY_SPACE>& mem, int& loc, long double& accum);
 
 //  Code = 32
 //  divide() => divide a word from a location in memory into the
 //  word in the accumulator (leaving result in accumulator)
-void divide(array<int, MEMORY_SPACE>& mem, int& loc, long double& accum);
+void divide(array<double, MEMORY_SPACE>& mem, int& loc, long double& accum);
 
 //  Code = 33
 //  multiply() => multiply a word from a location in memory by the 
 //  word in accumulator (leaving result in accumulator)
-void multiply(array<int, MEMORY_SPACE>& mem, int& loc, long double& accum);
+void multiply(array<double, MEMORY_SPACE>& mem, int& loc, long double& accum);
